Rework Linux Wi-Fi info retrieval as common Linux capability
