+++ lib/robots/common/linuxCapabilities/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+    LinuxWifiConfigurationCapability: require("./LinuxWifiConfigurationCapability"),
+};
