+++ lib/robots/roborock/capabilities/RoborockWifiConfigurationCapability.js
@@ -1,36 +1,41 @@
-const wireless = require("../../../utils/wireless");
-const WifiConfigurationCapability = require("../../../core/capabilities/WifiConfigurationCapability");
 const ValetudoWifiConfiguration = require("../../../entities/core/ValetudoWifiConfiguration");
+const LinuxWifiConfigurationCapability = require("../../common/linuxCapabilities/LinuxWifiConfigurationCapability");
 
-class RoborockWifiConfigurationCapability extends WifiConfigurationCapability {
+class RoborockWifiConfigurationCapability extends LinuxWifiConfigurationCapability {
+
+    getWifiInterface() {
+        return "wlan0";
+    }
+
     /**
      * @returns {Promise<ValetudoWifiConfiguration>}
      */
     async getWifiConfiguration() {
+        if (this.robot.config.get("embedded") === true) {
+            return await super.getWifiConfiguration();
+        }
+
         const output = {
             details: {
                 state: ValetudoWifiConfiguration.STATE.UNKNOWN
             }
         };
 
-        if (this.robot.config.get("embedded") === true) {
-            return wireless.getEmbeddedWirelessConfiguration("wlan0");
-        } else {
-            let res = await this.robot.sendCommand("get_network_info");
+        let res = await this.robot.sendCommand("get_network_info");
 
-            if (res !== "unknown_method") {
-                if (typeof res === "object" && res.bssid !== "") {
-                    output.details.state = ValetudoWifiConfiguration.STATE.CONNECTED;
+        if (res !== "unknown_method") {
+            if (typeof res === "object" && res.bssid !== "") {
+                output.details.state = ValetudoWifiConfiguration.STATE.CONNECTED;
 
-                    output.details.signal = parseInt(res.rssi);
-                    output.details.ips = [res.ip];
-                    output.ssid = res.ssid;
-                    output.details.frequency = ValetudoWifiConfiguration.FREQUENCY_TYPE.W2_4Ghz;
-                } else {
-                    output.details.state = ValetudoWifiConfiguration.STATE.NOT_CONNECTED;
-                }
+                output.details.signal = parseInt(res.rssi);
+                output.details.ips = [res.ip];
+                output.ssid = res.ssid;
+                output.details.frequency = ValetudoWifiConfiguration.FREQUENCY_TYPE.W2_4Ghz;
+            } else {
+                output.details.state = ValetudoWifiConfiguration.STATE.NOT_CONNECTED;
             }
         }
+
 
         return new ValetudoWifiConfiguration(output);
     }
